+++ src/main/java/com/telepathicgrunt/the_bumblezone/modCompat/ModChecker.java
@@ -20,10 +20,10 @@ public class ModChecker
 	 *
 	 * This will run the mod compat setup stuff. If it blows up, it attempts to catch the issue,
 	 * print it to the console, and then move on to the next mod instead of fully crashing. The compat
-	 * is basically optional and not neccessary for Bumblezone to function. If a classloading issue occurs
+	 * is basically optional and not necessary for Bumblezone to function. If a classloading issue occurs
 	 * somehow, we catch and print it so Forge doesn't silently swallow it. If this happens even with the
 	 * lambdas, at least it will be much easier to find and debug now although this breaks all mod compat
-	 * after the problematic mod lin.
+	 * after the problematic mod line.
 	 */
     public static void setupModCompat() {
 		String modid = "";